Create Laserphones
- Im SOOOO DUMB. I forgot that the C++ grid method is not like a normal coordinate point. in normal it is (x,y). In C++ is is (y,x) so i screwed up.
-also, since there are 2 C, you cant just end at a C.  Also, you cant say if they were not visited since they are all incremented when they +1;
  SOLUTION - just make the first -1 and then if the C value is not -1 do whatevah

Problem statement
Laserphones
===========

The cows have a new laser-based system so they can have casual
conversations while out in the pasture which is modeled as a W x H
grid of points (1 <= W <= 100; 1 <= H <= 100).

The system requires a sort of line-of-sight connectivity in order
to sustain communication. The pasture, of course, has rocks and
trees that disrupt the communication but the cows have purchased
diagonal mirrors that deflect the laser beam through a 90 degree
turn. Below is a map that illustrates the problem.

H is 8 and W is 7 for this map.  The two communicating cows are
notated as 'C's; rocks and other blocking elements are notated as
'*'s:

7 . . . . . . .         7 . . . . . . .
6 . . . . . . C         6 . . . . . /-C
5 . . . . . . *         5 . . . . . | *
4 * * * * * . *         4 * * * * * | *
3 . . . . * . .         3 . . . . * | .
2 . . . . . . .         2 . /-------/ .
1 . C . . . * .         1 . C . . . * .
0 . . . . . . .         0 . . . . . . .
  0 1 2 3 4 5 6           0 1 2 3 4 5 6

Determine the minimum number of mirrors M that must be installed to
maintain laser communication between the two cows.

PROBLEM NAME: lphone

INPUT FORMAT:

* Line 1: Two space separated integers: W and H

* Lines 2..H+1: The entire pasture.

SAMPLE INPUT:

7 8
.......
......C
......*
*****.*
....*..
.......
.C...*.
.......

OUTPUT FORMAT:

* Line 1: A single integer: M

SAMPLE OUTPUT:

3